add footer and move header files to new folder named layout inside component folder
