+++ src/app/page.tsx
@@ -1,11 +1,5 @@
-import FeedbackBox from '@/components/notices/FeedbackBox';
 import InspectionPage from '@/components/vc-inspection/InspectionPage';
 
 export default function Home() {
-  return (
-    <>
-      <InspectionPage />
-      <FeedbackBox />
-    </>
-  );
+  return <InspectionPage />;
 }
