+++ src/app/token/[id]/page.tsx
@@ -1,13 +1,7 @@
-import FeedbackBox from '@/components/notices/FeedbackBox';
 import InspectionPage from '@/components/vc-inspection/InspectionPage';
 
 export default async function Home({ params }: { params: { id: string } }) {
-  return (
-    <>
-      <InspectionPage defaultToken={base64urlDecode(params.id)} />
-      <FeedbackBox />
-    </>
-  );
+  return <InspectionPage defaultToken={base64urlDecode(params.id)} />;
 }
 
 function base64urlDecode(input: string): string {
