+++ src/components/notices/FeedbackBox.tsx
@@ -14,7 +14,7 @@ type FeedbackBoxProps = JSX.IntrinsicElements['div'] & {};
 export default function FeedbackBox({ className, ...props }: FeedbackBoxProps) {
   return (
     <div
-      className={cn('flex flex-col items-center justify-center gap-8 rounded-xl p-10  text-center', className)}
+      className={cn('flex flex-col items-center justify-center gap-2 rounded-xl p-10  text-center', className)}
       {...props}
     >
       <h2 className="text-2xl ">
