+++ src/stories/header/NavBar.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import NavBar from '@/components/header/NavBar';
+import NavBar from '@/components/layout/NavBar';
 
 const meta = {
   title: 'Header/NavBar',
